refactor: move config deployment to it's own file
